moved product editor into dialog

+++ src/app/components/product-card/product-card.component.ts
@@ -7,9 +7,10 @@ import {
   Output, Renderer2, ViewChild,
 } from '@angular/core';
 import {Product} from "../../models/product";
-import {MenuItem} from "primeng";
+import {ConfirmationService, MenuItem} from "primeng";
 import {ProductService} from "../../services/product.service";
 import {AuthenticationService} from "../../services/authentication.service";
+import {ProductEditorService} from "../../services/product-editor.service";
 
 @Component({
   selector: 'app-product-card',
@@ -30,7 +31,12 @@ export class ProductCardComponent implements OnInit {
 
   productContextItems: MenuItem[];
 
-  constructor(private productService: ProductService, public authService: AuthenticationService, private renderer2: Renderer2) {
+  constructor(private productService: ProductService,
+              public authService: AuthenticationService,
+              private renderer2: Renderer2,
+              private productEditorService: ProductEditorService,
+              private confirmationService: ConfirmationService
+  ) {
   }
 
   ngOnInit(): void {
@@ -41,15 +47,22 @@ export class ProductCardComponent implements OnInit {
       },
       {
         label: "Bearbeiten",
-        routerLink: "../admin/productEditor"
+        command: () => {
+          this.productEditorService.editProduct(this.product);
+        }
       },
       {
         label: "Löschen",
         command: () => {
-          this.productService.deleteProduct(this.product.id).subscribe(() => {
-            this.deleteProduct.emit(this.product.id);
-          },
-            (() => {}))
+          this.confirmationService.confirm({
+            message: 'Sind sie sicher, dass sie dieses Produkt löschen wolllen?',
+            accept: () => {
+              this.productService.deleteProduct(this.product.id).subscribe(() => {
+                this.deleteProduct.emit(this.product.id);
+              },
+                (error => {}))
+            }
+          });
         }
       }
     ];
